Fix for gem5 image conversion issue

+++ simulation/gem5-src/util/gem5img.py
@@ -352,7 +352,7 @@ newCom = Command(
 
 def newImage(file, mb):
     (cylinders, heads, sectors) = chsFromSize((mb * MB) / BlockSize)
-    size = cylinders * heads * sectors * BlockSize
+    size = int(cylinders * heads * sectors * BlockSize)
 
     # We lseek to the end of the file and only write one byte there. This
     # leaves a "hole" which many file systems are smart enough not to actually
